Fix text length

+++ packet.c
@@ -114,9 +114,9 @@ uint8_t make_t_pkt(char *buf, struct ctlr_cfg_t ctlr) {
 }
 
 #ifdef DEBUG
-void print_bytes(char *msg, uint8_t len) {
+void print_bytes(char *msg, uint16_t len) {
 	printf("(%s): length: %u, data:", __func__, len);
-	for (uint8_t i = 0; i < len; i++) {
+	for (uint16_t i = 0; i < len; i++) {
 		if (msg[i] >= 0x20 && msg[i] <= 0x7e) {
 			printf(" '%c'", msg[i]);
 		} else {
